+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ CHANGELOG
 Unreleased
 ==========
 
+* Changed: Support for scanner_id parameter for agents, agent-config, agent-exclusions, agent-groups, and
+bulk-operations API.
 * Added: Support for agent-exclusions API.
 * Added: Support for agent-config API.
 * Added: Support for bulk-operations API.